- UserServiceTest 클래스 testUserDeleteFail method 추가 및 성공!

+++ hahahohokiss/new-user-management/src/main/java/kr/ac/jejuuniv/service/UserServiceImpl.java
@@ -36,11 +36,17 @@ public class UserServiceImpl implements UserService {
 	}
 
 	public User modify(User user) {
-		return userRepository.update(user);
+		if(userRepository.findById(user.getId()) == null)
+			throw new DataNotFoundException();
+		else 
+			return userRepository.update(user);
 	}
 
 	public void delete(String id) {
-		userRepository.delete(id);
+		if(userRepository.findById(id) != null)
+			userRepository.delete(id);
+		else 
+			throw new DataNotFoundException();
 	}
 
 }
